+++ Services/NotificationService.cs
@@ -392,3 +392,14 @@ public class PushNotificationTemplate
 
     public bool IsChecked { get; set; }
 }
+
+public class NotificationRequestModel
+{
+    public string NotificationCode { get; set; }
+
+    public string NotificationTargetId { get; set; }
+
+    public string? ReceiverId { get; set; }
+
+    public string Message { get; set; }
+}
